BackEnd/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
boto3==1.34.32
botocore==1.34.149
CacheControl==0.14.0
cachetools==5.4.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.0
Django==4.2.9
django-celery==3.1.17
django-filter==24.3
django-multiselectfield==0.1.13
django-ses==4.1.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.17.0
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.62.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
idna==3.7
jmespath==1.0.1
kombu==5.3.7
msgpack==1.0.8
numpy==1.26.4
packaging==24.1
pandas==2.2.1
pillow==10.2.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.19
vine==5.1.0
wcwidth==0.2.13
